hashi_types::guardian

Module bitcoin_utils

Bitcoin utilities shared between Hashi and Guardian.

Two classes of types exist in this file:

Types with checked addresses that implement Serialize but not Deserialize

Types with unchecked addresses implement both Serialize and Deserialize

Structs

ExternalOutputUTXO

ExternalOutputUTXOWire

Copy of bitcoin_utils::ExternalOutputUTXO with unchecked address

InputUTXO

(Hashi+Guardian)-owned input UTXO TODO: Should we take derivation path as input instead of address & leaf_hash? Investigate later.

InputUTXOWire

Copy of bitcoin_utils::InputUTXO with unchecked address

InternalOutputUTXO

TxUTXOs

All the UTXOs associated with a withdrawal transaction

TxUTXOsWire

Copy of bitcoin_utils::TxUTXOs with unchecked address

Enums

OutputUTXO

Withdrawal destination and amount. External amounts count towards rate limits whereas internal amounts don’t. Internal address is derived inside the enclave to ensure that it is actually internal.

OutputUTXOWire

Copy of bitcoin_utils::OutputUTXO with unchecked address

Statics

BTC_LIB

Functions

compute_taproot_descriptor

Creates a taproot descriptor for the given enclave and hashi keys with a 2-of-2 multi_a script. Taproot addresses are constructed as follows:

construct_tx

Constructs a Bitcoin transaction with the given inputs and outputs.

get_derived_pubkey

Derives a child public key using unhardened derivation from a parent public key.

nums_internal_key

sign_btc_tx

Signs messages using Schnorr signatures (suitable for taproot script-spend).

single_key_taproot_script_path_address

single_key_taproot_script_path_spend_artifacts

Type Aliases

DerivationPath