hashi_types::proto::mpc_service_server

Trait MpcService

pub trait MpcService:
    Send
    + Sync
    + 'static {
    // Required methods
    fn send_messages<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SendMessagesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<SendMessagesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn retrieve_messages<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RetrieveMessagesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<RetrieveMessagesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn complain<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ComplainRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ComplainResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_public_mpc_output<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetPublicMpcOutputRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetPublicMpcOutputResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_reconfig_completion_signature<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetReconfigCompletionSignatureRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetReconfigCompletionSignatureResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_partial_signatures<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetPartialSignaturesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetPartialSignaturesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with MpcServiceServer.

Required Methods

fn send_messages<'life0, 'async_trait>( &'life0 self, request: Request<SendMessagesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<SendMessagesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Send dealer messages and receive signature.

fn retrieve_messages<'life0, 'async_trait>( &'life0 self, request: Request<RetrieveMessagesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<RetrieveMessagesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve a dealer’s messages.

fn complain<'life0, 'async_trait>( &'life0 self, request: Request<ComplainRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ComplainResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Send complaints about invalid shares and receive correct shares.

fn get_public_mpc_output<'life0, 'async_trait>( &'life0 self, request: Request<GetPublicMpcOutputRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetPublicMpcOutputResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get public MPC output.

fn get_reconfig_completion_signature<'life0, 'async_trait>( &'life0 self, request: Request<GetReconfigCompletionSignatureRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetReconfigCompletionSignatureResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get a validator’s signature on the reconfig completion message.

fn get_partial_signatures<'life0, 'async_trait>( &'life0 self, request: Request<GetPartialSignaturesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetPartialSignaturesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Pull partial signatures from another party.