hashi_types

Module move_types

Definitions of the raw Move structs in the hashi package

Structs

AbortReconfigEvent

Bag

Rust version of the Move sui::bag::Bag type.

BitcoinState

Rust version of the Move hashi::bitcoin_state::BitcoinState type.

BitcoinStateKey

Rust version of the Move hashi::bitcoin_state::BitcoinStateKey type.

BurnEvent

CertifiedMessage

Rust version of the Move hashi::committee::CertifiedMessage type.

Coin

CommittedRequestInfo

Lightweight info extracted from a request at commit time for validation.

Committee

This represents a BLS signing committee for a given epoch.

CommitteeMember

Rust version of the Move hashi::committee::CommitteeMember type.

CommitteeSet

Rust version of the Move hashi::committee_set::CommitteeSet type.

CommitteeSignature

Rust version of the Move hashi::committee::CommitteeSignature type.

Config

Rust version of the Move hashi::config::Config type.

DealerMessagesHashV1

Rust version of the Move hashi::tob::DealerMessagesHashV1 type.

DealerSubmissionV1

Rust version of the Move hashi::tob::DealerSubmissionV1 type.

DepositConfirmedEvent

DepositRequest

Rust version of the Move hashi::deposit_queue::DepositRequest type.

DepositRequestQueue

Rust version of the Move hashi::deposit_queue::DepositRequestQueue type.

DepositRequestedEvent

DisableVersion

Rust version of the Move hashi::disable_version::DisableVersion type.

EmergencyPause

Rust version of the Move hashi::emergency_pause::EmergencyPause type.

EnableVersion

Rust version of the Move hashi::enable_version::EnableVersion type.

EndReconfigEvent

Entry

Rust version of the Move sui::vec_map::Entry type.

EpochCertsV1

Rust version of the Move hashi::tob::EpochCertsV1 type.

ExpiredDepositDeletedEvent

Field

Hashi

Rust version of the Move hashi::hashi::Hashi type.

LinkedTable

Rust version of the Move sui::linked_table::LinkedTable type.

LinkedTableNode

Rust version of the Move sui::linked_table::Node type. This is the value stored in each dynamic field entry of a LinkedTable.

MemberInfo

Rust version of the Move hashi::committee_set::MemberInfo type.

MetadataCap

MintEvent

OutputUtxo

Rust version of the Move hashi::withdrawal_queue::OutputUtxo type.

PackageUpgradedEvent

Proposal

Rust version of the Move hashi::proposal::Proposal type.

ProposalCreatedEvent

ProposalDeletedEvent

ProposalExecutedEvent

QuorumReachedEvent

ReconfigCompletionMessage

Rust version of the Move struct hashi::reconfig::ReconfigCompletionMessage.

StartReconfigEvent

Table

Rust version of the Move sui::table::Table type (header only).

Treasury

Rust version of the Move hashi::treasury::Treasury type.

TreasuryCap

UpdateConfig

Rust version of the Move hashi::update_config::UpdateConfig type.

Upgrade

Rust version of the Move hashi::upgrade::Upgrade type.

UpgradeCap

Rust version of the Move sui::package::UpgradeCap type.

Utxo

UtxoId

txid:vout

UtxoPool

UtxoRecord

Rust version of the Move hashi::utxo_pool::UtxoRecord type.

UtxoSpentEvent

ValidatorRegistered

ValidatorUpdated

VecMap

Rust version of the Move sui::vec_map::VecMap type.

VecSet

Rust version of the Move sui::vec_set::VecSet type.

VoteCastEvent

VoteRemovedEvent

WithdrawalApprovedEvent

WithdrawalConfirmedEvent

WithdrawalPickedForProcessingEvent

WithdrawalRequest

Rust version of the Move hashi::withdrawal_queue::WithdrawalRequest type.

WithdrawalRequestQueue

Rust version of the Move hashi::withdrawal_queue::WithdrawalRequestQueue type.

WithdrawalRequestedEvent

WithdrawalSignedEvent

WithdrawalTransaction

Rust version of the Move hashi::withdrawal_queue::WithdrawalTransaction type.

Enums

ConfigValue

Rust version of the Move hashi::config_value::Value type.

HashiEvent

ProtocolType

Rust version of the Move hashi::tob::ProtocolType enum.

WithdrawalStatus

Rust version of the Move hashi::withdrawal_queue::WithdrawalStatus enum.

Traits

MoveType

Type Aliases

ObjectBag

Rust version of the Move sui::object_bag::ObjectBag type.