hashi_types::proto::guardian_service_server

Trait GuardianService

pub trait GuardianService:
    Send
    + Sync
    + 'static {
    // Required methods
    fn get_guardian_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetGuardianInfoRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetGuardianInfoResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn setup_new_key<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SetupNewKeyRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<SignedSetupNewKeyResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn operator_init<'life0, 'async_trait>(
        &'life0 self,
        request: Request<OperatorInitRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<OperatorInitResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn provisioner_init<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ProvisionerInitRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ProvisionerInitResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn standard_withdrawal<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SignedStandardWithdrawalRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<SignedStandardWithdrawalResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with GuardianServiceServer.

Required Methods

fn get_guardian_info<'life0, 'async_trait>( &'life0 self, request: Request<GetGuardianInfoRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetGuardianInfoResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Query the service for general information about its current state.

fn setup_new_key<'life0, 'async_trait>( &'life0 self, request: Request<SetupNewKeyRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<SignedSetupNewKeyResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Setup mode only: generate a new BTC key and distribute encrypted shares.

fn operator_init<'life0, 'async_trait>( &'life0 self, request: Request<OperatorInitRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<OperatorInitResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Operator initialization: provide config and commitments before provisioning.

fn provisioner_init<'life0, 'async_trait>( &'life0 self, request: Request<ProvisionerInitRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ProvisionerInitResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provisioner initialization: submit encrypted share and readiness info.

fn standard_withdrawal<'life0, 'async_trait>( &'life0 self, request: Request<SignedStandardWithdrawalRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<SignedStandardWithdrawalResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Standard withdrawal: request immediate withdrawal signature.